Moved helpers to utils subfolder
